requirements.txt
aiohttp==3.7.4.post0
aiohttp-jinja2==1.4.2
aioredis==1.3.0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
hiredis==2.0.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
PyYAML==5.4.1
trafaret==2.1.0
typing-extensions==3.10.0.0
yarl==1.6.3
